requirements.txt
nose
django
psycopg2
django-configurations
dj-database-url